:recycle: ディレクトリ配置構成を修正
