+++ src/app/components/spend/list/private/spend-private-list.component.scss
