+++ src/app/components/spend/list/public/spend-list.component.scss
