+++ src/app/components/spend/list/public/spend-list.component.ts
@@ -2,15 +2,16 @@ import { Component, ViewChild } from '@angular/core';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator, MatSort } from '@angular/material';
 
-import { SpendService } from '../../service/spend.service';
-import { Spend } from '../../model/spend';
+import { SpendService } from '../../../../service/spend.service';
+import { Spend } from '../../../../model/spend';
 
 @Component({
-  selector: 'app-spend-private-list',
-  templateUrl: './spend-private-list.component.html',
-  styleUrls: ['./spend-private-list.component.scss']
+  selector: 'app-spend-list',
+  templateUrl: './spend-list.component.html',
+  styleUrls: ['./spend-list.component.scss']
 })
-export class SpendPrivateListComponent {
+
+export class SpendListComponent {
   displayedColumns: string[] = ['createDate', 'category', 'amount'];
   dataSource: MatTableDataSource<Spend>;
   spendList: Spend[];
@@ -20,17 +21,13 @@ export class SpendPrivateListComponent {
   @ViewChild(MatSort) sort: MatSort;
 
   constructor(private spendService: SpendService) {
-    const data = this.spendService.getPrivateSpendList();
-
+    const data = this.spendService.getPublicSpendList();
 
     data.snapshotChanges().subscribe(item => {
       this.spendList = [];
       this.resultsLength = item.length;
-      console.log(item);
 
       item.reverse().forEach(element => {
-        console.log(element);
-
         const json = element.payload.toJSON();
         json['$key'] = element.key;
         this.spendList.push(json as Spend);
